ispravka tipa za cijenu iz double u int

+++ src/ba/unsa/etf/rpr/Korpa.java
@@ -19,9 +19,9 @@ public class Korpa {
     }
 
     public int dajUkupnuCijenuArtikala(){
-        double iznos = 0;
+        int iznos = 0;
         for(int i=0; i<=brElemenata; i++)
             iznos += artikli[i].getCijena();
-        return  iznos;
+        return iznos;
     }
 }
